+++ Changelog.md
@@ -1,11 +1,12 @@
 ### 0.4.13 (unreleased)
 
 Features:
+ * Syntax Checker: Deprecated "throw" in favour of require(), assert() and revert().
  * Type Checker: Warn if a local storage reference variable does not explicitly use the keyword ``storage``.
 
 Bugfixes:
- * Compiler Interface: Only output AST if analysis was successful.
  * Code Generator: Correctly unregister modifier variables.
+ * Compiler Interface: Only output AST if analysis was successful.
  * Error Output: Do not omit the error type.
 
 ### 0.4.12 (2017-07-03)